# swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "TRANSACTION CONTROL API.",
      "description": "This api is an api to create an app to control TRANSACTIONS / control expenses and income.",
      "contact": {
         "name": "Support",
         "email": "skp.tridimensional@gmail.com"
      },
      "version": "1.0.0"
   },
   "paths": {
      "/auth": {
         "get": {
            "tags": ["Auth"],
            "description": "Authentication rote.",
            "security": [
               {
                  "basicAuth": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Login successful.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ResponseAuth"
                        }
                     }
                  }
               },
               "404": {
                  "description": "When the User is not found in the database.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "status": {
                                 "type": "string",
                                 "example": "Error"
                              },
                              "message": {
                                 "type": "string",
                                 "example": "Authentication Failed"
                              }
                           }
                        }
                     }
                  }
               },
               "401": {
                  "description": "When User Wrong Password.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "status": {
                                 "type": "string",
                                 "example": "Error"
                              },
                              "message": {
                                 "type": "string",
                                 "example": "Email or password Is Incorrect!"
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/transaction": {
         "post": {
            "tags": ["Transactions"],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "description": "Create of Transaction.",
            "requestBody": {
               "description": "Create a new Transaction passing the following data...",
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/RequestTransaction"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Created realized with success.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ResponseTransaction"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Invalid data."
               },
               "401": {
                  "description": "Token is missing."
               }
            }
         },
         "get": {
            "parameters": [
               {
                  "in": "query",
                  "name": "month",
                  "type": "string",
                  "description": "Get by Month of create."
               }
            ],
            "tags": ["Transactions"],
            "description": "Get transaction by months orderBy(CREATE DATE).",
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "responses": {
               "200": {
                  "description": "Success",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ResponseListTransactions"
                        }
                     }
                  }
               },
               "404": {
                  "description": "Errors 400",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "status": {
                                 "type": "string",
                                 "example": "Error"
                              },
                              "message": {
                                 "type": "string",
                                 "example": "Just Filter By Month between 1 and 12 42 IsInvalid Mount"
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/transaction/{transaction_id}": {
         "delete": {
            "parameters": [
               {
                  "in": "path",
                  "name": "transaction_id",
                  "type": "integer",
                  "required": true
               }
            ],
            "tags": ["Transactions"],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "description": "Delete a transaction.",
            "responses": {
               "204": {},
               "400": {}
            }
         }
      },
      "/user/": {},
      "/user/profile": {
         "get": {
            "tags": ["Profile"],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "description": "Configuration your profile for first time.",
            "responses": {
               "200": {
                  "description": "Created realized with success.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ResponseGetProfile200"
                        }
                     }
                  }
               },
               "404": {
                  "description": "The User already has a registered profile.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ResponseGetProfile400"
                        }
                     }
                  }
               },
               "401": {
                  "description": "Token is missing."
               }
            }
         },
         "post": {
            "tags": ["Profile"],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "description": "Configuration your profile for first time.",
            "requestBody": {
               "description": "Create a Profile passing the following data...",
               "content": {
                  "multipart/form-data": {
                     "schema": {
                        "$ref": "#/components/schemas/RequestProfile"
                     }
                  }
               }
            },
            "responses": {
               "201": {
                  "description": "Created realized with success."
               },
               "400": {
                  "description": "The User already has a registered profile.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ResponseCreateProfile400"
                        }
                     }
                  }
               },

               "401": {
                  "description": "Token is missing."
               }
            }
         },
         "patch": {
            "tags": ["Profile"],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "description": "Updated your profile settings.",
            "requestBody": {
               "description": "Updated a Profile passing the following data...",
               "content": {
                  "multipart/form-data": {
                     "schema": {
                        "$ref": "#/components/schemas/RequestProfile"
                     }
                  }
               }
            },
            "responses": {
               "201": {
                  "description": "Created realized with success."
               },
               "400": {
                  "description": "The User already has a registered profile.",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ResponseUpdatedProfile400"
                        }
                     }
                  }
               },

               "401": {
                  "description": "Token is missing."
               }
            }
         }
      }
   },
   "components": {
      "schemas": {
         "ResponseGetProfile400": {
            "type": "object",
            "properties": {
               "status": {
                  "type": "string",
                  "example": "Error"
               },
               "message": {
                  "type": "string",
                  "description": "When user don't have an profile register.",
                  "example": "Profile not found"
               }
            }
         },
         "ResponseGetProfile200": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "string",
                  "format": "uuid"
               },
               "avatar": {
                  "type": "string",
                  "description": "Url obtida do firebase storage para a imagem do avatar.",
                  "example": "https://storage.googleapis.com/control-cb6dd.appspot.com/images/user-%3F943488fb-ad60-48ff-8eda-13b0963abb92?GoogleAccessId=firebase-adminsdk-cnxr9%40control-cb6dd.iam.gserviceaccount.com&Expires=1742180400&Signature=eIAO4s9FJh3%2F9giFvLKQfW%2FTnnZOPq3y66PAaYaifz74ELUNXMA%2F%2FtteVKI%2FGt7j91FXnBUy9WZg%2Br7CnyDxRyeENy8AVNTWE5mSCFljk0xOUb79XZyMxAXAEPlFjfbcR%2BpxdhRug8goxpI5XBQzchMOK%2FAheQOthaSlSZyFueWg7yxgSMY8mABogdlvZ5I4UFGEyod9csaV6gp5MyVqw5ZaFI2Qayx%2BSp8FKXJcmVeHPaW5HwcZ%2BprxIljkqQRcpvZokQP%2F78LHXtGiFeSPdxtjeuIgNJGJWN8ghmxpmm966Nb4IvTF1D%2Bdrsy5oceicwtLVu79xj%2BEFJbvxCDUFQ%3D%3D"
               },
               "profession": {
                  "type": "string",
                  "description": "Profission.",
                  "example": "dev-backend"
               },
               "salary": {
                  "type": "string",
                  "description": "Salario.",
                  "example": "1008.99"
               },
               "phonenumber": {
                  "type": "string",
                  "description": "PhoneNumber.",
                  "example": "(11) 99999-9999"
               },
               "dateOfBirth": {
                  "type": "string",
                  "description": "Data de aniversario.",
                  "example": "20/06/2002"
               },
               "user": {
                  "type": "object",
                  "description": "Devolve uma obj contendo id,email e nome do usuário.",
                  "properties": {
                     "name": {
                        "type": "string",
                        "example": "joão"
                     },
                     "email": {
                        "type": "string",
                        "example": "joão@gmail.com"
                     },
                     "id": {
                        "type": "string",
                        "format": "uuid"
                     }
                  }
               }
            }
         },
         "RequestTransaction": {
            "type": "object",
            "properties": {
               "value": {
                  "type": "string",
                  "description": "Value of the Transaction to be created (IN HEXADECIMAL).",
                  "example": "-422232.00"
               },
               "description": {
                  "type": "string",
                  "description": "A description of what this transaction is about.",
                  "example": "New clothes"
               },
               "categoryType": {
                  "enum": [
                     "transport",
                     "food",
                     "habitation",
                     "education",
                     "health",
                     "leisure",
                     "products",
                     "debts",
                     "Taxes",
                     "Investments",
                     "unknown"
                  ],
                  "description": "Enter a category for this Transaction.",
                  "example": "products"
               },
               "dueDate": {
                  "type": "string",
                  "description": "(EXPENSES ONLY) - Enter a due date for this transaction.",
                  "example": "2023-06-11"
               },
               "filingDate": {
                  "type": "string",
                  "description": "(REVENUE ONLY) - Enter a receipt date for this transaction",
                  "example": "2023-06-11"
               }
            }
         },
         "ResponseTransaction": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "string",
                  "description": "uuid id generated to reference the created transaction",
                  "format": "uuid"
               },
               "value": {
                  "type": "string",
                  "description": "Value of the Transaction to be created (IN HEXADECIMAL).",
                  "example": "-422232.00"
               },
               "description": {
                  "type": "string",
                  "description": "A description of what this transaction is about.",
                  "example": "New clothes"
               },
               "categoryType": {
                  "enum": [
                     "transport",
                     "food",
                     "habitation",
                     "education",
                     "health",
                     "leisure",
                     "products",
                     "debts",
                     "Taxes",
                     "Investments",
                     "unknown"
                  ],
                  "description": "Enter a category for this Transaction.",
                  "example": "products"
               },
               "dueDate": {
                  "type": "string",
                  "description": "(EXPENSES ONLY) - Enter a due date for this transaction.",
                  "example": "2023-06-11"
               },
               "filingDate": {
                  "type": "string",
                  "description": "(REVENUE ONLY) - Enter a receipt date for this transaction",
                  "example": "2023-06-11"
               }
            }
         },
         "ResponseListTransactions": {
            "type": "object",
            "properties": {
               "monthBalense": {
                  "type": "object",
                  "properties": {
                     "expense": {
                        "type": "string",
                        "description": "Value of the expenses found referring to the transactions found.",
                        "example": "-422232.00"
                     },
                     "revenue": {
                        "type": "string",
                        "description": "Value of the revenues found referring to the transactions found.",
                        "example": "-2550515.33"
                     },
                     "total": {
                        "type": "string",
                        "description": "Total amount of revenues/expenses found referring to the transactions found.",
                        "example": "-2550515.33"
                     }
                  }
               },
               "transactions": {
                  "type": "array",
                  "description": "List of transactions found",
                  "items": {
                     "type": "object",
                     "properties": {
                        "id": {
                           "type": "string",
                           "description": "Id(uuid) unique identification referring to the transaction.",
                           "format": "uuid"
                        },
                        "description": {
                           "type": "string",
                           "description": "Transaction Description.",
                           "example": "New clothes"
                        },
                        "value": {
                           "type": "string",
                           "description": "Transaction Value.",
                           "example": "-422232.00"
                        },
                        "recurrence": {
                           "type": ["string", "null"],
                           "description": "Informs which recurrence the transaction has, if not, it returns null.",
                           "example": "null"
                        },
                        "installments": {
                           "type": ["string", "null"],
                           "description": "Informs the number of installments the transaction contains, or returns null if it is a recipe.",
                           "example": "null"
                        },
                        "isSubscription": {
                           "type": ["boolean", "null"],
                           "description": "Change a true case if it is an expense and if it is an enrollment, if it is not an enrollment it returns false, if it is an income it returns null.",
                           "example": "false"
                        },
                        "due_date": {
                           "type": ["string", "null"],
                           "description": "Returns the due date of the expense in question, if it is not an expense, it returns null.",
                           "example": "2023-06-11T03:00:00.000Z"
                        },
                        "filingDate": {
                           "type": ["string", "null"],
                           "description": "Returns the receipt date of the recipe in question, if it is not a recipe, it returns null.",
                           "example": "false"
                        },
                        "resolved": {
                           "type": "string",
                           "description": "Returns the status of the transaction if it has been completed or not.",
                           "example": "2023-04-24T19:53:15.704Z"
                        },
                        "created_at": {
                           "type": "string",
                           "description": "Indicates the creation date of the transaction in question.",
                           "example": "2023-04-24T19:53:15.704Z"
                        },
                        "updated_at": {
                           "type": "string",
                           "description": "Indicates the update date of the transaction in question.",
                           "example": "2023-04-24T19:53:15.704Z"
                        },
                        "type": {
                           "type": "string",
                           "description": "Indicates the type of transaction whether it is an income or an expense.",
                           "example": "expense"
                        },
                        "userId": {
                           "type": "string",
                           "description": "Indicates the ID of the client that owns the transaction.",
                           "format": "uuid"
                        },
                        "transactionsCategoryId": {
                           "type": "string",
                           "description": "Indicates the transaction id.",
                           "format": "uuid"
                        },
                        "category": {
                           "type": "object",
                           "description": "Indicates the category in which the transaction belongs, by default the category is 'unknown'",
                           "properties": {
                              "id": {
                                 "type": "string",
                                 "description": "Indicates the transaction id.",
                                 "format": "uuid"
                              },
                              "name": {
                                 "type": "string",
                                 "description": "Indicates the transaction name.",
                                 "format": "uuid"
                              },
                              "created_at": {
                                 "type": "string",
                                 "description": "Indicates the creation date of the category in question.",
                                 "example": "2023-04-24T19:53:15.704Z"
                              },
                              "updated_at": {
                                 "type": "string",
                                 "description": "Indicates the update date of the category in question.",
                                 "example": "2023-04-24T19:53:15.704Z"
                              }
                           }
                        }
                     }
                  }
               }
            }
         },
         "ResponseAuth": {
            "type": "object",
            "properties": {
               "user": {
                  "type": "object",
                  "properties": {
                     "name": {
                        "type": "string",
                        "description": "User name that just authenticated.",
                        "example": "Leandro Santos"
                     },
                     "email": {
                        "type": "string",
                        "description": "Email of the user who has just authenticated.",
                        "example": "leandro@example.com"
                     }
                  }
               },
               "token": {
                  "type": "string",
                  "description": "Token JWT para autenticação.",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImxlYW5kcm9AZ21haWwuY29tIiwiaWF0IjoxNjgyMzc3MDI1LCJleHAiOjE2ODIzODA2MjUsInN1YiI6ImE4YmQ2NDQyLWFjNjktNDFjNy04ZjZmLTgxM2UyM2M0MGUzZiJ9.kFa_H7L93v7HNSL6HvBtQOpGF4RgK48CiHA-9A2w0L8"
               },
               "PushNotificationToken": {
                  "type": "string",
                  "description": "JWT token for push notification (not implemented yet).",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImxlYW5kcm9AZ21haWwuY29tIiwiaWF0IjoxNjgyMzc3MDI1LCJleHAiOjE2ODIzODA2MjUsInN1YiI6ImE4YmQ2NDQyLWFjNjktNDFjNy04ZjZmLTgxM2UyM2M0MGUzZiJ9.kFa_H7L93v7HNSL6HvBtQOpGF4RgK48CiHA-9A2w0L8"
               }
            }
         },
         "RequestProfile": {
            "type": "object",
            "description": "Profile Creation.",
            "properties": {
               "props": {
                  "type": "object",
                  "description": "Enter the data to create/updated the profile.",
                  "properties": {
                     "Birthday": {
                        "type": "string",
                        "description": "User's birthday (Must be over 16 years old).",
                        "example": "20/06/2000"
                     },
                     "phonenumber": {
                        "type": "string",
                        "description": "User's phone number. (Must follow the following format) (optional).",
                        "example": "(11) 99999-9999"
                     },
                     "profession": {
                        "type": "string",
                        "description": "Enter the Profession in which you work (optional).",
                        "example": "Software Engineer - Back-end."
                     },
                     "salary": {
                        "type": "string",
                        "description": "Enter the user's salary/income in hexadecimal (optional).",
                        "example": "1245.22"
                     },
                     "profile_id": {
                        "type": "string",
                        "description": "Enter the user's profile id. (only on patch request).",
                        "format": "uuid",
                        "required": false
                     }
                  }
               },
               "avatar": {
                  "type": "string",
                  "description": "Choose a profile picture (optional).",
                  "format": "binary"
               }
            }
         },
         "ResponseCreateProfile400": {
            "type": "object",
            "properties": {
               "status": {
                  "type": "string",
                  "example": "Error"
               },
               "message": {
                  "type": "string",
                  "description": "The User already has a registered profile.",
                  "example": "User already has a profile"
               }
            }
         },
         "ResponseUpdatedProfile400": {
            "type": "string",
            "example": "err"
         }
      },
      "securitySchemes": {
         "bearerAuth": {
            "description": "Authenticate using JWT",
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         },
         "basicAuth": {
            "description": "Login Authentication.",
            "type": "http",
            "scheme": "basic"
         }
      }
   }
}